Fixed rounding error bug in servingsMultiplier

+++ app/src/main/java/com/gerardbradshaw/mater/activities/recipedetail/RecipeDetailActivity.java
@@ -214,7 +214,6 @@ public class RecipeDetailActivity extends AppCompatActivity {
   }
 
   private void updateServings() {
-
     // Set up dialog for user confirmation
     AlertDialog.Builder alertBuilder = new AlertDialog.Builder(this);
     final EditText input = new EditText(this);
@@ -228,7 +227,7 @@ public class RecipeDetailActivity extends AppCompatActivity {
       public void onClick(DialogInterface dialogInterface, int i) {
 
         customServings = Integer.parseInt(input.getText().toString());
-        double servingsMultiplier = customServings / defaultServings;
+        double servingsMultiplier = (double) customServings / defaultServings;
 
         // Reset customIngredients
         customIngredients.clear();
